Import FatFS-0.10a, koji izgleda ima bolji automounter za removable media, i dopusta proizvoljno imenovanje "diskova", koji se sad zovu "C:" i "D:" umjesto "0:" i "1:".
git-svn-id: svn+ssh://login.nxlab.fer.hr/fpgasvn/f32c@1891 6164e1d0-f123-df11-80f4-0019d13f85ef

+++ src/lib/diskio.c
@@ -229,7 +229,7 @@ disk_write(BYTE drive, const BYTE* Buffer, DWORD SectorNumber, BYTE SectorCount)
 DRESULT
 disk_ioctl(BYTE drive, BYTE cmd, void* buf)
 {
-	uint32_t *up = buf;
+	WORD *up = buf;
 
 	switch (cmd) {
 	case GET_SECTOR_SIZE:
